scripts/requirements.txt
pillow==12.0.0
typer==0.20.0
